+++ src/src/zKillSubscriber.ts
@@ -1,13 +1,14 @@
-import { Client, Colors } from 'discord.js';
+import { Client } from 'discord.js';
 import { MessageEvent, WebSocket } from 'ws';
 import { REST } from '@discordjs/rest';
 import AsyncLock from 'async-lock';
 import { Queue, Worker } from 'bullmq';
-import { v4 as uuidv4 } from 'uuid';
 import * as fs from 'fs';
 import { EsiClient } from './lib/esiClient';
-import { sendKillMailToDiscord } from './lib/sendKill';
+import { sendKillMailToDiscord } from './jobs/sendKill';
 import { initializeQueueDashboard } from './lib/queueDashboard';
+import { IZkill } from './interfaces/zkill';
+import processKill from './jobs/processKill';
 
 export enum SubscriptionType {
     ALL = 'all',
@@ -42,7 +43,7 @@ interface SubscriptionChannel {
     subscriptions: Map<string, Subscription>
 }
 
-interface Subscription {
+export interface Subscription {
     subType: SubscriptionType
     id?: number
     minValue: number,
@@ -51,63 +52,56 @@ interface Subscription {
     killType?: KillType
 }
 
-export interface SolarSystem {
-    id: number
-    regionId: number
-    regionName: string
-    constellationId: number
-    constellationName: string
-}
-
 export class ZKillSubscriber {
     protected static instance: ZKillSubscriber;
     protected doClient: Client;
 
     protected subscriptions: Map<string, SubscriptionGuild>;
-    protected systems: Map<number, SolarSystem>;
     protected ships: Map<number, number>;
     protected rest: REST;
 
     protected asyncLock: AsyncLock;
     protected esiClient: EsiClient;
 
-    protected worker? : Worker;
-    protected queue? : Queue;
+    protected worker: Worker;
+    protected queue: Queue;
 
     protected constructor(client: Client) {
         this.asyncLock = new AsyncLock({ maxPending: 10000 });
         this.esiClient = new EsiClient();
         this.subscriptions = new Map<string, SubscriptionGuild>();
-        this.systems = new Map<number, SolarSystem>();
         this.ships = new Map<number, number>();
         this.loadConfig();
-        this.loadSystems();
         this.loadShips();
 
-        if (process.env.USE_REDIS == 'true') {
-            // Initialize the worker
-            this.worker = new Worker('zkillboard', async (job) => {
-                const data = job.data;
+        // Initialize the worker
+        this.worker = new Worker('zkillboard', async (job) => {
+            const data = job.data;
+            if (job.name === 'process-kill') {
+                await processKill(this, this.queue, job);
+            }
+            else {
                 await sendKillMailToDiscord(this, data.guildId, data.channelId, data.subType, data.data, data.subId, data.messageColor);
-            }, {
-                connection: {
-                    host: 'redis',
-                    port: 6379,
-                },
-                concurrency: 10,
-            });
+            }
+        }, {
+            connection: {
+                host: 'redis',
+                port: 6379,
+            },
+            concurrency: 10,
+        });
 
-            // Initialize the queue
-            this.queue = new Queue('zkillboard', {
-                connection: {
-                    host: 'redis',
-                    port: 6379,
-                },
-            });
+        // Initialize the queue
+        this.queue = new Queue('zkillboard', {
+            connection: {
+                host: 'redis',
+                port: 6379,
+            },
+        });
 
-            // initialize the dashboard
-            initializeQueueDashboard(this.queue);
-        }
+        // initialize the dashboard
+        initializeQueueDashboard(this.queue);
+
 
         this.doClient = client;
         this.rest = new REST({ version: '9' }).setToken(process.env.DISCORD_BOT_TOKEN || '');
@@ -137,188 +131,26 @@ export class ZKillSubscriber {
     }
 
     protected async onMessage(event: MessageEvent) {
-        const data = JSON.parse(event.data.toString());
-        this.subscriptions.forEach((guild, guildId) => {
-            guild.channels.forEach((channel, channelId) => {
-                channel.subscriptions.forEach(async (subscription) => {
-                    let color: number = Colors.Green;
-                    try {
-                        let requireSend = false, systemData = null, groupId = null;
-                        if (subscription.minValue > data.zkb.totalValue) {
-                            return; // Do not send if below the min value
-                        }
-                        switch (subscription.subType) {
-                        case SubscriptionType.PUBLIC:
-                            await this.sendKill(guildId, channelId, subscription.subType, data);
-                            break;
-                        case SubscriptionType.REGION:
-                            systemData = await this.getSystemData(data.solar_system_id);
-                            if (systemData.regionId === subscription.id) {
-                                await this.sendKill(guildId, channelId, subscription.subType, data);
-                            }
-                            break;
-                        case SubscriptionType.CONSTELLATION:
-                            systemData = await this.getSystemData(data.solar_system_id);
-                            if (systemData.constellationId === subscription.id) {
-                                await this.sendKill(guildId, channelId, subscription.subType, data);
-                            }
-                            break;
-                        case SubscriptionType.SYSTEM:
-                            if (data.solar_system_id === subscription.id) {
-                                await this.sendKill(guildId, channelId, subscription.subType, data);
-                            }
-                            break;
-                        case SubscriptionType.ALLIANCE:
-                            if (data.victim.alliance_id === subscription.id) {
-                                requireSend = subscription.killType === KillType.LOSSES || subscription.killType === undefined;
-                                color = Colors.Red;
-                            }
-                            if (!requireSend) {
-                                for (const attacker of data.attackers) {
-                                    if (attacker.alliance_id === subscription.id) {
-                                        requireSend = subscription.killType === KillType.KILLS || subscription.killType === undefined;
-                                        break;
-                                    }
-                                }
-                            }
-                            if (requireSend) {
-                                if (subscription.limitType !== LimitType.NONE && !await this.isInLimit(subscription, data.solar_system_id)) {
-                                    return;
-                                }
-                                await this.sendKill(guildId, channelId, subscription.subType, data, subscription.id, color);
-                            }
-                            break;
-                        case SubscriptionType.corporation:
-                            if (data.victim.corporation_id === subscription.id) {
-                                requireSend = subscription.killType === KillType.LOSSES || subscription.killType === undefined;
-                                color = Colors.Red;
-                            }
-                            if (!requireSend) {
-                                for (const attacker of data.attackers) {
-                                    if (attacker.corporation_id === subscription.id) {
-                                        requireSend = subscription.killType === KillType.KILLS || subscription.killType === undefined;
-                                        break;
-                                    }
-                                }
-                            }
-                            if (requireSend) {
-                                if (subscription.limitType !== LimitType.NONE && !await this.isInLimit(subscription, data.solar_system_id)) {
-                                    return;
-                                }
-                                await this.sendKill(guildId, channelId, subscription.subType, data, subscription.id, color);
-                            }
-                            break;
-                        case SubscriptionType.CHARACTER:
-                            if (data.victim.character_id === subscription.id) {
-                                requireSend = subscription.killType === KillType.LOSSES || subscription.killType === undefined;
-                                color = Colors.Red;
-                            }
-                            if (!requireSend) {
-                                for (const attacker of data.attackers) {
-                                    if (attacker.character_id === subscription.id) {
-                                        requireSend = subscription.killType === KillType.KILLS || subscription.killType === undefined;
-                                        break;
-                                    }
-                                }
-                            }
-                            if (requireSend) {
-                                if (subscription.limitType !== LimitType.NONE && !await this.isInLimit(subscription, data.solar_system_id)) {
-                                    return;
-                                }
-                                await this.sendKill(guildId, channelId, subscription.subType, data, subscription.id, color);
-                            }
-                            break;
-                        case SubscriptionType.GROUP:
-                            if (data.victim.ship_type_id) {
-                                groupId = await this.getShipGroup(data.victim.ship_type_id);
-                                if (groupId === subscription.id) {
-                                    requireSend = subscription.killType === KillType.LOSSES || subscription.killType === undefined;
-                                    color = Colors.Red;
-                                }
-                            }
-                            if (!requireSend) {
-                                for (const attacker of data.attackers) {
-                                    if (attacker.ship_type_id) {
-                                        groupId = await this.getShipGroup(attacker.ship_type_id);
-                                        if (groupId === subscription.id) {
-                                            requireSend = subscription.killType === KillType.KILLS || subscription.killType === undefined;
-                                            break;
-                                        }
-                                    }
-                                }
-                            }
-                            if (requireSend) {
-                                if (subscription.limitType !== LimitType.NONE && !await this.isInLimit(subscription, data.solar_system_id)) {
-                                    return;
-                                }
-                                await this.sendKill(guildId, channelId, subscription.subType, data, subscription.id, color);
-                            }
-                            break;
-                        case SubscriptionType.SHIP:
-                            if (data.victim.ship_type_id) {
-                                if (data.victim.ship_type_id === subscription.id) {
-                                    requireSend = subscription.killType === KillType.LOSSES || subscription.killType === undefined;
-                                    color = Colors.Red;
-                                }
-                            }
-                            if (!requireSend) {
-                                for (const attacker of data.attackers) {
-                                    if (attacker.ship_type_id) {
-                                        if (attacker.ship_type_id === subscription.id) {
-                                            requireSend = subscription.killType === KillType.KILLS || subscription.killType === undefined;
-                                            break;
-                                        }
-                                    }
-                                }
-                            }
-                            if (requireSend) {
-                                if (subscription.limitType !== LimitType.NONE && !await this.isInLimit(subscription, data.solar_system_id)) {
-                                    return;
-                                }
-                                await this.sendKill(guildId, channelId, subscription.subType, data, subscription.id, color);
-                            }
-                            break;
-                        default:
-                        }
-                    }
-                    catch (e) {
-                        console.log(e);
-                    }
-                });
-            });
-        });
-    }
+        const data : IZkill = JSON.parse(event.data.toString());
 
-    private async sendKill(guildId: string, channelId: string, subType: SubscriptionType, data: any, subId?: number, messageColor: number = Colors.Grey) {
-        if (this.queue) {
-            await this.queue.add(
-                'zkillboard',
-                { guildId, channelId, subType, data, subId, messageColor },
-                {
-                    jobId: uuidv4(),
-                    removeOnComplete: {
-                        age: 300,
-                    }, // Automatically remove the job after 30 seconds
-                    backoff: {
-                        type: 'fixed',
-                        delay: 60000, // Wait 1 minute before retrying
-                    },
-                    attempts: 10, // Retry 10 times
-                    removeOnFail: {
-                        age: 86400,
-                    }, // Remove after 24 hours
+        // Add the kill to queue for processing
+        this.queue.add(
+            'process-kill',
+            data,
+            {
+                jobId: 'kill-' + data.killmail_id.toString(),
+                removeOnComplete: true,
+                backoff: {
+                    type: 'fixed',
+                    delay: 60000, // Wait 1 minute before retrying
                 },
-            );
-            return;
-        }
-
-        await this.asyncLock.acquire('sendKill', async (done) => {
-            await sendKillMailToDiscord(this, guildId, channelId, subType, data, subId, messageColor);
-            done();
-        });
-
+                attempts: 10, // Retry 10 times
+                removeOnFail: {
+                    age: 86400,
+                }, // Remove after 24 hours
+            },
+        );
     }
-
     public static getInstance(client?: Client) {
         if (!this.instance && client) {
             this.instance = new ZKillSubscriber(client);
@@ -431,48 +263,6 @@ export class ZKillSubscriber {
         return map;
     }
 
-    private async getSystemData(systemId: number): Promise<SolarSystem> {
-        return await this.asyncLock.acquire('fetchSystem', async (done) => {
-
-            let system = this.systems.get(systemId);
-            if (system) {
-                done(undefined, system);
-                return;
-            }
-            system = await this.esiClient.getSystemInfo(systemId);
-            this.systems.set(systemId, system);
-            fs.writeFileSync('./config/systems.json', JSON.stringify(Object.fromEntries(this.systems)), 'utf8');
-
-            done(undefined, system);
-        });
-    }
-
-    private async getShipGroup(shipId: number): Promise<number> {
-        return await this.asyncLock.acquire('fetchShip', async (done) => {
-
-            let group = this.ships.get(shipId);
-            if (group) {
-                done(undefined, group);
-                return;
-            }
-            group = await this.esiClient.getTypeGroupId(shipId);
-            this.ships.set(shipId, group);
-            fs.writeFileSync('./config/ships.json', JSON.stringify(Object.fromEntries(this.ships)), 'utf8');
-
-            done(undefined, group);
-        });
-    }
-
-    private loadSystems() {
-        if (fs.existsSync('./config/systems.json')) {
-            const fileContent = fs.readFileSync('./config/systems.json', 'utf8');
-            const data = JSON.parse(fileContent);
-            for (const key in data) {
-                this.systems.set(Number.parseInt(key), data[key] as SolarSystem);
-            }
-        }
-    }
-
     private loadShips() {
         if (fs.existsSync('./config/ships.json')) {
             const fileContent = fs.readFileSync('./config/ships.json', 'utf8');
@@ -483,20 +273,12 @@ export class ZKillSubscriber {
         }
     }
 
-    private async isInLimit(subscription: Subscription, solar_system_id: number) {
-        const systemData = await this.getSystemData(solar_system_id);
-        const limit = subscription.limitIds?.split(',') || [];
-        if (subscription.limitType === LimitType.SYSTEM && limit.indexOf(systemData.id.toString()) !== -1) {
-            return true;
-        }
-        if (subscription.limitType === LimitType.CONSTELLATION && limit.indexOf(systemData.constellationId.toString()) !== -1) {
-            return true;
-        }
-        return subscription.limitType === LimitType.REGION && limit.indexOf(systemData.regionId.toString()) !== -1;
+    getGuildSubscriptions(guildId: string): SubscriptionGuild | undefined {
+        return this.subscriptions.get(guildId);
     }
 
-    getSubscriptions(guildId: string): SubscriptionGuild | undefined {
-        return this.subscriptions.get(guildId);
+    getAllSubscriptions(): Map<string, SubscriptionGuild> {
+        return this.subscriptions;
     }
 
     getDoClient(): Client {